+++ components/api_server/src/utils/type.py
@@ -7,6 +7,10 @@ from typing import Literal, NewType, TypedDict, TYPE_CHECKING
 if TYPE_CHECKING:
     from datetime import datetime
 
+    from model.report import Report
+    from shared.model.metric import Metric
+    from shared.model.source import Source
+    from shared.model.subject import Subject
 
 Change = dict[str, str | dict[str, str]]
 EditScope = Literal["source", "metric", "subject", "report", "reports"]
@@ -47,3 +51,13 @@ class SessionData(TypedDict, total=False):
     email: str
     common_name: str
     session_expiration_datetime: datetime
+
+
+@dataclass
+class SourceContext:
+    """Source, combined with its containing metric, subject, and report."""
+
+    metric: Metric
+    report: Report
+    source: Source
+    subject: Subject