+++ docs/src/changelog.md
@@ -15,6 +15,7 @@ If your currently installed *Quality-time* version is not the penultimate versio
 ### Fixed
 
 - When creating an issue from Quality-time in Jira, make sure that the link back to Quality-time doesn't expand any items but only goes to the metric. Fixes [#12481](https://github.com/ICTU/quality-time/issues/12481).
+- When mass editing a source parameter, only change sources that actually have the parameter being changed. Fixes [#12482](https://github.com/ICTU/quality-time/issues/12482).
 - Count a Jira issue linked to multiple metrics only once in the issues dashboard card. Fixes [#12483](https://github.com/ICTU/quality-time/issues/12483).
 - Remove restart policy from docker-compose.yml. Fixes [#12506](https://github.com/ICTU/quality-time/issues/12506).
 - Fix missing service in renderer Helm chart that was causing PDF-exports to fail on Kubernetes. Fixes [#12514](https://github.com/ICTU/quality-time/issues/12514).